+++ aion-lang-intellij/src/main/kotlin/com/aion/intellij/AionFile.kt
@@ -1,0 +1,9 @@
+package com.aion.intellij
+
+import com.intellij.extapi.psi.PsiFileBase
+import com.intellij.psi.FileViewProvider
+
+class AionFile(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, AionLanguage) {
+    override fun getFileType() = AionFileType
+    override fun toString() = "Aion File"
+}