+++ aion-lang-intellij/src/main/kotlin/com/aion/intellij/AionLanguage.kt
@@ -1,0 +1,7 @@
+package com.aion.intellij
+
+import com.intellij.lang.Language
+
+object AionLanguage : Language("Aion") {
+    private fun readResolve(): Any = AionLanguage
+}