+++ aion-lang-intellij/src/main/kotlin/com/aion/intellij/AionSyntaxHighlighterFactory.kt
@@ -1,0 +1,11 @@
+package com.aion.intellij
+
+import com.intellij.openapi.fileTypes.SyntaxHighlighter
+import com.intellij.openapi.fileTypes.SyntaxHighlighterFactory
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.vfs.VirtualFile
+
+class AionSyntaxHighlighterFactory : SyntaxHighlighterFactory() {
+    override fun getSyntaxHighlighter(project: Project?, virtualFile: VirtualFile?): SyntaxHighlighter =
+        AionSyntaxHighlighter()
+}